requirements.txt
yfinance>=0.2.38
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pyotp>=2.9.0
logzero>=1.7.0
smartapi-python>=1.3.5
openai>=1.12.0
duckduckgo-search>=5.3.0
flask>=3.0.0
flask-cors>=4.0.0
feedparser>=6.0.10
vaderSentiment>=3.3.2
transformers>=4.30.0
torch>=2.0.0
